+++ RELEASE-NOTES.md
@@ -7,6 +7,8 @@
 - **Feature** - Add demo lessons #2
 - **Model** - Create video collection data models #8
 - **Model** - Add `restricted_permit` attribute to device data model #9
+- **Feature** - Only permitted devices can access restricted videos #10
+- **Feature** - Add VB demo lessons, Y-GRE demo lessons, and Y-GRE test reviews #11
 
 **IMPROVED**
 - **Model** - Change `demo` attribute of video data model to `restricted` #7